Revert "chore: Harden against async blocking deadlocks (#27653)"
This reverts commit 0d00d44b2962c91d375b8a4fdd054ad46f0d87b1.
